third_party/WebKit/LayoutTests/media/video-src-invalid-remove.html

<!DOCTYPE html>
<title>Test that removing "src" attribute does not trigger load of "source" elements.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");

    video.onloadedmetadata = t.unreached_func();
    video.onerror = t.step_func(function() {
        assert_equals(video.error.code, MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED);

        setTimeout(t.step_func_done(), 200);
        video.removeAttribute("src");
    });

    video.src = "bogus.mov";
    var source = document.createElement("source");
    source.src = "content/test.ogv";
    video.appendChild(source);
});
</script>
